Renaming add delete frame by manage words frame
