+++ frames/manage_words_frame.py
